% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr-package.R
\docType{data}
\name{.datatable.aware}
\alias{.datatable.aware}
\title{dtplyr is data.table aware}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\description{
dtplyr is data.table aware
}
\keyword{internal}
