% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr.R
\name{dt_as}
\alias{dt_as}
\title{Make a new data.table}
\usage{
dt_as(df)
}
\arguments{
\item{df}{An R object to coerce to a data.table}
}
\value{
data.table
}
\description{
Create a new data.table by copying the underlying data and coercing
it to the data.table type. This is a synonym for as.data.table.
}
\examples{
dt_as(data.frame(a = 1:5))

}
\seealso{
\code{\link[data.table]{as.data.table}}
}
