% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr.R
\name{dt_merge}
\alias{dt_merge}
\title{Use data.table merge routines to join two tables}
\usage{
dt_merge(df, ...)
}
\arguments{
\item{df}{An R object that can be coerced to a data.table}

\item{...}{Additional arguments for merge}
}
\value{
data.table
}
\description{
Coerce the first input to a data.table, and then hand off to the
higher performance merge routine.
}
\examples{
lh <- data.frame(a = 1:5, b = letters[1:5])
rh <- data.frame(a = 2:4, c = letters[24:26])
dt_merge(lh, rh)
\dontrun{
lh \%>\% dt_merge(rh)
}

}
\seealso{
\code{\link[data.table]{merge}}
}
