% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr.R
\name{dt_op}
\alias{dt_op}
\title{Run an operation on a data.table}
\usage{
dt_op(df, ...)
}
\arguments{
\item{df}{An R object that can be coerced to a data.table}

\item{...}{Additional arguments for [.data.table}
}
\value{
data.table
}
\description{
Expose data.table's square-bracket-based functionality as a method.
This makes it more natural to use with the pipe operator.
}
\examples{
df <- data.frame(a = c(1, 1, 2, 2, 3, 3), b = 1:6)
dt_op(df, a > 1, sum(b), keyby = a)
df \%>\% dt_op(a > 1, sum(b), keyby = a)

}
\seealso{
\code{\link[data.table]{data.table}}
}
