% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr-package.R
\docType{package}
\name{dtplyr-package}
\alias{dtplyr}
\alias{dtplyr-package}
\title{dtplyr: Data Table Back-End for 'dplyr'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides a data.table backend for 'dplyr'. The goal
    of 'dtplyr' is to allow you to write 'dplyr' code that is
    automatically translated to the equivalent, but usually much faster,
    data.table code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverse/dtplyr}
  \item Report bugs at \url{https://github.com/tidyverse/dtplyr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
