% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-by.R
\name{key_by}
\alias{key_by}
\title{Group by with ordering}
\usage{
key_by(.data, ...)
}
\arguments{
\item{.data}{A [lazy_dt]}

\item{...}{Variables to key by}
}
\description{
Uses [data.table::setkey()]; this guarantees that the output is sorted
and typically increases performance when peforming multiple operations on
the same grouping.
}
\examples{
library(dplyr)
library(data.table)

mtcars2 <- lazy_dt(mtcars)

# group_by() doesn't order by group
mtcars2 \%>\% group_by(cyl) \%>\% summarise(mpg = mean(mpg))

# key_by() does
mtcars2 \%>\% key_by(cyl) \%>\% summarise(mpg = mean(mpg))
}
