% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-first.R
\name{lazy_dt}
\alias{lazy_dt}
\title{Create a "lazy" data.table}
\usage{
lazy_dt(x, name = NULL, immutable = TRUE)
}
\arguments{
\item{x}{A data table (or something can can be coerced to a data table)}

\item{name}{Optionally, supply a name to be used in generated expressions.
For expert use only.}

\item{immutable}{If `TRUE`, `x` is treated as immutable and will never
be modified in place. Alternatively, set `immutable = FALSE` to state
that it's ok for dtplyr to modify the input.}
}
\description{
A lazy data.table lazy captures the intent of dplyr verbs, only actually
performing computation when requested (with [connect()], [as.data.frame()],
[data.table::as.data.table()], or [tibble::as_tibble()]). This allows dtplyr
to convert dplyr verbs into as few data.table expressions as possible, which
leads to a high performance translation.

See `vignette("translation")` for more details.
}
\examples{
library(dplyr)

mtcars2 <- lazy_dt(mtcars)
mtcars2
mtcars2 \%>\% select(mpg:cyl)
mtcars2 \%>\% select(x = mpg, y = cyl)
mtcars2 \%>\% filter(cyl == 4) \%>\% select(mpg)
mtcars2 \%>\% select(mpg, cyl) \%>\% filter(cyl == 4)
mtcars2 \%>\% mutate(cyl2 = cyl * 2, cyl4 = cyl2 * 2)
mtcars2 \%>\% transmute(cyl2 = cyl * 2, vs2 = vs * 2)
mtcars2 \%>\% filter(cyl == 8) \%>\% mutate(cyl2 = cyl * 2)

by_cyl <- mtcars2 \%>\% group_by(cyl)
by_cyl \%>\% summarise(mpg = mean(mpg))
by_cyl \%>\% mutate(mpg = mean(mpg))
by_cyl \%>\% filter(mpg < mean(mpg)) \%>\% summarise(hp = mean(hp))
}
