% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R, R/step-group.R
\name{single_table}
\alias{single_table}
\alias{group_by.dtplyr_step}
\title{Single table operations}
\usage{
\method{group_by}{dtplyr_step}(.data, ..., add = FALSE, arrange = TRUE)
}
\arguments{
\item{.data}{a tbl}

\item{...}{Variables to group by. All tbls accept variable names.
Some tbls will accept functions of variables. Duplicated groups
will be silently dropped.}

\item{add}{When \code{add = FALSE}, the default, \code{group_by()} will
override existing groups. To add to the existing groups, use
\code{add = TRUE}.}

\item{arrange}{If \code{TRUE}, will automatically arrange the output of
subsequent grouped operations by group. If \code{FALSE}, output order will be
left unchanged. In the generated data.table code this switches between
using the \code{keyby} (\code{TRUE}) and \code{by} (\code{FALSE}) arguments.}
}
\description{
This documents differences between standard dplyr verbs and their
data.table instantiation.
}
