% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-subset.R
\name{slice.dtplyr_step}
\alias{slice.dtplyr_step}
\title{Subset rows using their positions}
\usage{
\method{slice}{dtplyr_step}(.data, ...)
}
\arguments{
\item{.data}{A \code{\link[=lazy_dt]{lazy_dt()}}.}

\item{...}{Positive integers giving rows to select, or negative
integers giving rows to drop.}
}
\description{
This is a method for the dplyr \code{\link[=slice]{slice()}} generic. It is translated to
the \code{i} argument of \verb{[.data.table}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

dt <- lazy_dt(mtcars)
dt \%>\% slice(1, 5, 10)

dt \%>\%
  group_by(cyl) \%>\%
  slice(1)

dt \%>\%
  group_by(cyl) \%>\%
  slice(-1)
}
