% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src_dt}
\alias{src_dt}
\title{A local data table source.}
\usage{
src_dt(pkg = NULL, env = NULL)
}
\arguments{
\item{pkg, env}{Either the name of a package or an environment object in
which to look for objects.}
}
\description{
This is mainly useful for testing, since makes it possible to refer to
local and remote tables using exactly the same syntax.
}
