% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtplyr-package.R
\docType{package}
\name{dtplyr-package}
\alias{dtplyr}
\alias{dtplyr-package}
\title{dtplyr: Data Table Back-End for 'dplyr'}
\description{
This implements the data table back-end for 'dplyr'
    so that you can seamlessly use data table and 'dplyr' together.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverse/dtplyr}
  \item Report bugs at \url{https://github.com/tidyverse/dtplyr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Authors:
\itemize{
  \item Lionel Henry \email{lionel@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
